+++ CHANGELOG.rst
@@ -9,14 +9,19 @@ Changelog
 0.0.15
 ------
 * Add 'example'
+* Disable converting urls in "full" settings
+* Fix filebrowser error
 
-0.0.13
+0.0.14
 ------
 * Add default tinymce content CSS styles
 * Add eslint
 * Remove unused grappelli and filebrowser js files
 * Update to TinyMCE 4.3.4
 
+0.0.13
+------
+* Update to TinyMCE 4.3.3
 
 0.0.12
 ------